(Refactor) Combine SearchDetailFollower.js and Follow.js
