+++ src/pages/Follow.js
@@ -1,23 +1,53 @@
-import React, { useEffect, useState } from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
+import { useRecoilValue, useSetRecoilState } from 'recoil';
 import { useHistory, useLocation, NavLink } from 'react-router-dom';
-import { useRecoilValue } from 'recoil';
 import styled from 'styled-components';
 
-import { myFollowListByType } from '../recoil/states/user';
+import { refreshSearchUserState } from '../recoil/states/search';
 
 import { MonsterListItem } from '../components/monster';
 import { BackButtonHeader } from '../components/common';
 import { Gnb } from '../components/gnb';
 
-const FollowPage = () => {
+import { userApis } from '../api';
+import { OK } from '../constants/statusCode';
+
+import { disappearScrollbar } from '../styles/Mixin';
+
+const Follow = () => {
   const history = useHistory();
   const location = useLocation();
   const [followList, setFollowList] = useState(null);
+
+  const userMonsterCode = location.pathname.split('/')[2];
+
   const tabType = location?.search?.split('tab=')?.[1];
   const isFollowTab = tabType === 'followers' || tabType === 'following';
-  const getFollowList = useRecoilValue(myFollowListByType(tabType));
-  const goToMyPage = () => history.push('mypage/information');
   const isActiveTab = (type) => tabType === type;
+
+  const refreshSearchUserInfo = useSetRecoilState(refreshSearchUserState);
+
+  const getUserList = useCallback(async () => {
+    if (!isFollowTab) {
+      return;
+    }
+
+    await setFollowList(null);
+
+    let getUserResponse =
+      tabType === 'followers'
+        ? userApis.getUserFollowers(userMonsterCode)
+        : userApis.getUserFollowings(userMonsterCode);
+
+    const { data } = await getUserResponse;
+
+    if (data.statusCode === OK) {
+      const followList =
+        tabType === 'followers' ? data.followers : data.followings;
+
+      setFollowList(followList ?? []);
+    }
+  }, [tabType, isFollowTab, userMonsterCode]);
 
   useEffect(() => {
     if (!isFollowTab) {
@@ -26,22 +56,30 @@ const FollowPage = () => {
   }, [history, tabType, isFollowTab]);
 
   useEffect(() => {
-    setFollowList(getFollowList);
+    getUserList();
+  }, [getUserList]);
 
-    return () => {
-      setFollowList(null);
-    };
-  }, [getFollowList]);
+  const onClickgoBack = () => {
+    refreshSearchUserInfo((id) => id + 1);
+
+    if (history.length <= 2) {
+      history.replace(`/`);
+    }
+    history.goBack();
+  };
 
   return (
     <>
       <FollowContainer>
-        <BackButtonHeader onButtonClick={goToMyPage} marginBottom="0" />
+        <BackButtonHeader onButtonClick={onClickgoBack} />
         <NavButtonWrap>
           <NavButtonItem>
             <NavButton
               isActive={() => isActiveTab('followers')}
-              to="/follow?tab=followers"
+              to={{
+                pathname: `/follow/${userMonsterCode}`,
+                search: `?tab=followers`,
+              }}
               activeClassName="active"
             >
               팔로워
@@ -50,7 +88,10 @@ const FollowPage = () => {
           <NavButtonItem>
             <NavButton
               isActive={() => isActiveTab('following')}
-              to="/follow?tab=following"
+              to={{
+                pathname: `/follow/${userMonsterCode}`,
+                search: `?tab=following`,
+              }}
               activeClassName="active"
             >
               팔로잉
@@ -89,8 +130,6 @@ const FollowPage = () => {
     </>
   );
 };
-
-export default FollowPage;
 
 const FollowContainer = styled.div`
   background-color: var(--bg-wrapper);
@@ -138,10 +177,11 @@ const NavButton = styled(NavLink)`
   }
 `;
 
-const FollowListWrap = styled.div`
+const FollowListWrap = styled.ul`
   height: calc(100% - 120px);
   padding-top: 16px;
   overflow-y: auto;
+  ${disappearScrollbar()};
 `;
 
 const FollowList = styled.ul`
@@ -163,3 +203,5 @@ const EmptyPlace = styled.div`
     font-weight: var(--weight-semi-regular);
   }
 `;
+
+export default Follow;
